chore: harmonize millesime in db & allow flexible query param
2019-2020, 2020_2019...
